Clarify method of joint control
Now it explicitly sets the control force of the named robot joints.

It also chooses a joint effort at random and sticks with it for the
whole run. So random but boring movement.

Ball shooters can also set their force to 0, so you can use them
to zero-out control effort after the first frame.