Add hit_the_gym to demonstrate gym wrapper
Hopefully this is useful in running some learning.

Has high and low dimensional state options. And the cameras that are
part of the high-d state are listed by name here.

Note that the ball shooter still needs to be in the loop because
it needs to give the ball velocity in the first timestep. This
might need to be improved -- baked into the HitBallEnv somewhere?