Make episodes 256 timesteps instead of 300
For one, 300 was too many. It was 10 seconds. The
slowest speed I have used for the random speed range
was 0.5 m/s. So it could travel at least 5m. And the
arm is between 2.5m and 3m from the ball at the start.

And 256 is divisible by the default 128 rollout run.
Which just fits nicer with the way it trains. This way
episodes always line up with the 128-step status.