Fix range of depth to be useful
Mujoco normalizes from 0 to 1, but since we have no background,
it was including its own max distance of 145m in that. This
squashes all our relevant depths down to nothing.

In the old version, it was also still 0-1... and then we were converting
to uint8, so all 0. Now we make it 0-255 (or perhaps 0-254, but
whatever).