Give reward for returning ball to +x direction
Triggered after it as crossed within 2m of robot, then
gets reward for distance beyond 2m the ball is in +x.

This might mitigate the failed contact detection? If it gets
reward for it bouncing off anything in the right direction.

And it will encourage it hitting things in the more impressive
direction; more clean contact.