Detect contact made between control timesteps
Not all ball contacts last long enough to be currently in contact
at a control timestep, and previous contact detection only worked
for that moment.

Instead, we keep track of contacts made in each mujoco timestep.
And reset at each new sim step. And print things out that are
helpful to understand the scope of the problem.